Enemy Damage/Knockback
The enemy can now damage the player when they run into them, this will cause both of them to be knocked back a bit to give the player a lil space to recover.
The player can now also damage the enemy with the hammer and kill them.

If the enemy has more than 1 HP then the enemy will be knocked back by the attack so the player can't just insta-kill them with 1 attack.