Boss 1 & Final Boss
Boss 1 has a script now, it summons holes in the ground after every other hit, basic but it works.
The final boss also has a script, it has a shield around it that can only be broken by the burst from the three artifacts. Once the shield is broken it is vulnerable for 2 hits, after which it reforms it's shield and summons 2 minions. Repeat this for another two hits and it'll summon 4 minions. Then just break it one more time and defeat it, and you win!